+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed
 - Fix styles for inline mode broken by previous wrong rebases
+- Remove `blocks.json`
 
 ### Changed
 - Adapt to use choiceType on assembly options as enums (SINGLE, TOGGLE, MULTIPLE)